+++ src/main/java/net/grid/vampiresdelight/data/builder/package-info.java
@@ -1,0 +1,7 @@
+@MethodsReturnNonnullByDefault
+@ParametersAreNonnullByDefault
+package net.grid.vampiresdelight.data.builder;
+
+import net.minecraft.MethodsReturnNonnullByDefault;
+
+import javax.annotation.ParametersAreNonnullByDefault;