+++ src/main/java/net/grid/vampiresdelight/data/recipe/VDCookingRecipes.java
@@ -3,24 +3,17 @@ package net.grid.vampiresdelight.data.recipe;
 import de.teamlapen.vampirism.core.ModBlocks;
 import de.teamlapen.vampirism.core.ModItems;
 import de.teamlapen.vampirism.core.ModTags;
-import net.grid.vampiresdelight.VampiresDelight;
 import net.grid.vampiresdelight.common.registry.VDItems;
 import net.grid.vampiresdelight.common.tag.VDCommonTags;
+import net.grid.vampiresdelight.data.builder.VDCookingPotRecipeBuilder;
 import net.minecraft.data.recipes.RecipeOutput;
-import net.minecraft.resources.ResourceLocation;
-import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.Items;
 import net.minecraft.world.item.crafting.Ingredient;
-import net.minecraft.world.level.block.Block;
 import net.neoforged.neoforge.common.Tags;
 import vectorwing.farmersdelight.client.recipebook.CookingPotRecipeBookTab;
-import vectorwing.farmersdelight.data.builder.CookingPotRecipeBuilder;
 
 import java.util.stream.Stream;
-
-import static net.grid.vampiresdelight.common.utility.VDNameUtils.blockName;
-import static net.grid.vampiresdelight.common.utility.VDNameUtils.itemName;
 
 @SuppressWarnings("unused")
 public class VDCookingRecipes {
@@ -41,23 +34,23 @@ public class VDCookingRecipes {
     }
 
     private static void cookMiscellaneous(RecipeOutput output) {
-        CookingPotRecipeBuilder.cookingPotRecipe(VDItems.DAISY_TEA.get(), 1, NORMAL_COOKING, MEDIUM_EXP)
+        VDCookingPotRecipeBuilder.cookingPotRecipe(VDItems.DAISY_TEA.get(), 1, NORMAL_COOKING, MEDIUM_EXP)
                 .addIngredient(Items.OXEYE_DAISY)
                 .addIngredient(Items.OXEYE_DAISY)
                 .addIngredient(Items.SUGAR)
                 .unlockedByAnyIngredient(Items.OXEYE_DAISY)
                 .setRecipeBookTab(CookingPotRecipeBookTab.DRINKS)
-                .build(output, itemLocationCooking(VDItems.DAISY_TEA.get()));
+                .build(output);
 
-        CookingPotRecipeBuilder.cookingPotRecipe(VDItems.ORCHID_TEA.get(), 1, NORMAL_COOKING, MEDIUM_EXP)
+        VDCookingPotRecipeBuilder.cookingPotRecipe(VDItems.ORCHID_TEA.get(), 1, NORMAL_COOKING, MEDIUM_EXP)
                 .addIngredient(VDCommonTags.FOODS_MILK)
                 .addIngredient(VDItems.ORCHID_PETALS.get())
                 .addIngredient(VDItems.ORCHID_PETALS.get())
                 .unlockedByAnyIngredient(VDItems.ORCHID_PETALS.get(), ModBlocks.VAMPIRE_ORCHID.get())
                 .setRecipeBookTab(CookingPotRecipeBookTab.DRINKS)
-                .build(output, itemLocationCooking(VDItems.ORCHID_TEA.get()));
+                .build(output);
 
-        CookingPotRecipeBuilder.cookingPotRecipe(VDItems.WEIRD_JELLY_BLOCK.get(), 1, NORMAL_COOKING, MEDIUM_EXP)
+        VDCookingPotRecipeBuilder.cookingPotRecipe(VDItems.WEIRD_JELLY_BLOCK.get(), 1, NORMAL_COOKING, MEDIUM_EXP)
                 .addIngredient(Tags.Items.BONES)
                 .addIngredient(Tags.Items.SLIMEBALLS)
                 .addIngredient(Tags.Items.SLIMEBALLS)
@@ -66,18 +59,18 @@ public class VDCookingRecipes {
                 .addIngredient(ModTags.Items.PURE_BLOOD)
                 .unlockedByAnyIngredient(ModItems.PURE_BLOOD_0.get(), ModItems.PURE_BLOOD_1.get(), ModItems.PURE_BLOOD_2.get(), ModItems.PURE_BLOOD_3.get(), ModItems.PURE_BLOOD_4.get())
                 .setRecipeBookTab(CookingPotRecipeBookTab.MEALS)
-                .build(output, itemLocationCooking(VDItems.WEIRD_JELLY_BLOCK.get()));
+                .build(output);
 
-        CookingPotRecipeBuilder.cookingPotRecipe(VDItems.TRICOLOR_DANGO.get(), 1, NORMAL_COOKING, MEDIUM_EXP, Items.STICK)
+        VDCookingPotRecipeBuilder.cookingPotRecipe(VDItems.TRICOLOR_DANGO.get(), 1, NORMAL_COOKING, MEDIUM_EXP, Items.STICK)
                 .addIngredient(VDCommonTags.FOODS_DOUGH_RICE)
                 .addIngredient(VDItems.BLOOD_SYRUP.get())
                 .addIngredient(VDItems.ORCHID_PETALS.get())
                 .addIngredient(Items.SUGAR)
                 .unlockedByAnyIngredient(VDItems.ORCHID_PETALS.get(), VDItems.BLOOD_SYRUP.get())
                 .setRecipeBookTab(CookingPotRecipeBookTab.MISC)
-                .build(output, itemLocationCooking(VDItems.TRICOLOR_DANGO.get()));
+                .build(output);
 
-        CookingPotRecipeBuilder.cookingPotRecipe(VDItems.MULLED_WINE_GLASS.get(), 1, NORMAL_COOKING, MEDIUM_EXP)
+        VDCookingPotRecipeBuilder.cookingPotRecipe(VDItems.MULLED_WINE_GLASS.get(), 1, NORMAL_COOKING, MEDIUM_EXP)
                 .addIngredient(VDItems.BLOOD_WINE_GLASS.get())
                 .addIngredient(Ingredient.fromValues(Stream.of(
                         new Ingredient.ItemValue(new ItemStack(ModBlocks.CURSED_ROOTS.get())),
@@ -88,11 +81,11 @@ public class VDCookingRecipes {
                 .addIngredient(Items.SUGAR)
                 .unlockedByAnyIngredient(VDItems.BLOOD_WINE_GLASS.get())
                 .setRecipeBookTab(CookingPotRecipeBookTab.DRINKS)
-                .build(output, itemLocationCooking(VDItems.MULLED_WINE_GLASS.get()));
+                .build(output);
     }
 
     private static void cookMeals(RecipeOutput output) {
-        CookingPotRecipeBuilder.cookingPotRecipe(VDItems.ORCHID_CREAM_SOUP.get(), 1, NORMAL_COOKING, MEDIUM_EXP)
+        VDCookingPotRecipeBuilder.cookingPotRecipe(VDItems.ORCHID_CREAM_SOUP.get(), 1, NORMAL_COOKING, MEDIUM_EXP)
                 .addIngredient(VDItems.ORCHID_PETALS.get())
                 .addIngredient(Tags.Items.FOODS_VEGETABLE)
                 .addIngredient(VDCommonTags.CROPS_ONION)
@@ -100,18 +93,18 @@ public class VDCookingRecipes {
                 .addIngredient(VDCommonTags.FOODS_MILK)
                 .unlockedByAnyIngredient(VDItems.ORCHID_PETALS.get())
                 .setRecipeBookTab(CookingPotRecipeBookTab.MEALS)
-                .build(output, itemLocationCooking(VDItems.ORCHID_CREAM_SOUP.get()));
+                .build(output);
 
-        CookingPotRecipeBuilder.cookingPotRecipe(VDItems.BLACK_MUSHROOM_SOUP.get(), 1, NORMAL_COOKING, MEDIUM_EXP)
+        VDCookingPotRecipeBuilder.cookingPotRecipe(VDItems.BLACK_MUSHROOM_SOUP.get(), 1, NORMAL_COOKING, MEDIUM_EXP)
                 .addIngredient(VDItems.BLACK_MUSHROOM.get())
                 .addIngredient(Items.POTATO)
                 .addIngredient(Items.CARROT)
                 .addIngredient(VDCommonTags.CROPS_ONION)
                 .unlockedByAnyIngredient(VDItems.BLACK_MUSHROOM.get())
                 .setRecipeBookTab(CookingPotRecipeBookTab.MEALS)
-                .build(output, itemLocationCooking(VDItems.BLACK_MUSHROOM_SOUP.get()));
+                .build(output);
 
-        CookingPotRecipeBuilder.cookingPotRecipe(VDItems.ORCHID_CURRY.get(), 1, NORMAL_COOKING, MEDIUM_EXP)
+        VDCookingPotRecipeBuilder.cookingPotRecipe(VDItems.ORCHID_CURRY.get(), 1, NORMAL_COOKING, MEDIUM_EXP)
                 .addIngredient(VDItems.ORCHID_PETALS.get())
                 .addIngredient(VDItems.ORCHID_PETALS.get())
                 .addIngredient(Items.POTATO)
@@ -122,9 +115,9 @@ public class VDCookingRecipes {
                 )))
                 .unlockedByAnyIngredient(VDItems.ORCHID_PETALS.get())
                 .setRecipeBookTab(CookingPotRecipeBookTab.MEALS)
-                .build(output, itemLocationCooking(VDItems.ORCHID_CURRY.get()));
+                .build(output);
 
-        CookingPotRecipeBuilder.cookingPotRecipe(VDItems.BLACK_MUSHROOM_NOODLES.get(), 1, NORMAL_COOKING, MEDIUM_EXP)
+        VDCookingPotRecipeBuilder.cookingPotRecipe(VDItems.BLACK_MUSHROOM_NOODLES.get(), 1, NORMAL_COOKING, MEDIUM_EXP)
                 .addIngredient(VDItems.BLACK_MUSHROOM.get())
                 .addIngredient(VDItems.BLACK_MUSHROOM.get())
                 .addIngredient(VDCommonTags.FOODS_PASTA)
@@ -132,17 +125,17 @@ public class VDCookingRecipes {
                 .addIngredient(VDCommonTags.FOODS_MILK)
                 .unlockedByAnyIngredient(VDItems.BLACK_MUSHROOM.get())
                 .setRecipeBookTab(CookingPotRecipeBookTab.MEALS)
-                .build(output, itemLocationCooking(VDItems.BLACK_MUSHROOM_NOODLES.get()));
+                .build(output);
 
-        CookingPotRecipeBuilder.cookingPotRecipe(VDItems.GARLIC_SOUP.get(), 1, NORMAL_COOKING, MEDIUM_EXP, Items.BREAD)
+        VDCookingPotRecipeBuilder.cookingPotRecipe(VDItems.GARLIC_SOUP.get(), 1, NORMAL_COOKING, MEDIUM_EXP, Items.BREAD)
                 .addIngredient(VDCommonTags.FOODS_RAW_CHICKEN)
                 .addIngredient(VDCommonTags.CROPS_GARLIC)
                 .addIngredient(Tags.Items.FOODS_VEGETABLE)
                 .unlockedByAnyIngredient(VDItems.ROASTED_GARLIC.get(), ModBlocks.GARLIC.asItem())
                 .setRecipeBookTab(CookingPotRecipeBookTab.MEALS)
-                .build(output, itemLocationCooking(VDItems.GARLIC_SOUP.get()));
+                .build(output);
 
-        CookingPotRecipeBuilder.cookingPotRecipe(VDItems.BORSCHT.get(), 1, NORMAL_COOKING, MEDIUM_EXP)
+        VDCookingPotRecipeBuilder.cookingPotRecipe(VDItems.BORSCHT.get(), 1, NORMAL_COOKING, MEDIUM_EXP)
                 .addIngredient(Ingredient.fromValues(Stream.of(
                         new Ingredient.TagValue(VDCommonTags.FOODS_RAW_PORK),
                         new Ingredient.ItemValue(new ItemStack(Items.BEEF)),
@@ -156,12 +149,12 @@ public class VDCookingRecipes {
                 .addIngredient(VDCommonTags.CROPS_GARLIC)
                 .unlockedByAnyIngredient(VDItems.ROASTED_GARLIC.get(), ModBlocks.GARLIC.asItem())
                 .setRecipeBookTab(CookingPotRecipeBookTab.MEALS)
-                .build(output, itemLocationCooking(VDItems.BORSCHT.get()));
+                .build(output);
     }
 
     // Temporary
     private static void fermentingAlternatives(RecipeOutput output) {
-        CookingPotRecipeBuilder.cookingPotRecipe(VDItems.DANDELION_BEER_BOTTLE.get(), 1, SLOW_FERMENTING, LARGE_EXP, Items.GLASS_BOTTLE)
+        VDCookingPotRecipeBuilder.cookingPotRecipe(VDItems.DANDELION_BEER_BOTTLE.get(), 1, SLOW_FERMENTING, LARGE_EXP, Items.GLASS_BOTTLE)
                 .addIngredient(Items.WHEAT)
                 .addIngredient(Items.WHEAT)
                 .addIngredient(Items.WHEAT)
@@ -169,9 +162,9 @@ public class VDCookingRecipes {
                 .addIngredient(Items.SUGAR)
                 .unlockedByAnyIngredient(Items.WHEAT, Items.DANDELION)
                 .setRecipeBookTab(CookingPotRecipeBookTab.DRINKS)
-                .build(output, itemLocationCooking(VDItems.DANDELION_BEER_BOTTLE.get()));
+                .build(output);
 
-        CookingPotRecipeBuilder.cookingPotRecipe(VDItems.BLOOD_WINE_BOTTLE.get(), 1, SLOW_FERMENTING, LARGE_EXP, Items.GLASS_BOTTLE)
+        VDCookingPotRecipeBuilder.cookingPotRecipe(VDItems.BLOOD_WINE_BOTTLE.get(), 1, SLOW_FERMENTING, LARGE_EXP, Items.GLASS_BOTTLE)
                 .addIngredient(VDItems.BLOOD_SYRUP.get())
                 .addIngredient(VDItems.BLOOD_SYRUP.get())
                 .addIngredient(VDItems.BLOOD_SYRUP.get())
@@ -179,14 +172,6 @@ public class VDCookingRecipes {
                 .addIngredient(Items.SUGAR)
                 .unlockedByAnyIngredient(VDItems.BLOOD_SYRUP.get())
                 .setRecipeBookTab(CookingPotRecipeBookTab.DRINKS)
-                .build(output, itemLocationCooking(VDItems.BLOOD_WINE_BOTTLE.get()));
-    }
-
-    private static String itemLocationCooking(Item item) {
-        return ResourceLocation.fromNamespaceAndPath(VampiresDelight.MODID, "cooking/" + itemName(item)).toString();
-    }
-
-    private static String blockLocationCooking(Block block) {
-        return ResourceLocation.fromNamespaceAndPath(VampiresDelight.MODID, "cooking/" + blockName(block)).toString();
+                .build(output);
     }
 }